[hsmtool] Rename the spxef module
Rename the `spxef` module to `extra::spxef` as we'll be adding more
custom backends into the `extra` module.
