+++ sw/host/hsmtool/BUILD
@@ -44,7 +44,8 @@ rust_library(
         "src/lib.rs",
         "src/module.rs",
         "src/profile.rs",
-        "src/spxef/mod.rs",
+        "src/extra/mod.rs",
+        "src/extra/spxef.rs",
         "src/util/attribute/attr.rs",
         "src/util/attribute/attribute_type.rs",
         "src/util/attribute/certificate_type.rs",
