+++ sw/host/hsmtool/src/extra/mod.rs
@@ -2,10 +2,6 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#![feature(min_specialization)]
-pub mod commands;
-pub mod error;
-pub mod extra;
-pub mod module;
-pub mod profile;
-pub mod util;
+mod spxef;
+
+pub use spxef::SpxEf;
