+++ sw/host/hsmtool/src/extra/spxef.rs
